minor changes, and made it compile on mac os x

+++ lib/new.cpp
@@ -1,11 +1,11 @@
 #include <liballoc.h>
 
-void* operator new(unsigned int size)
+void* operator new(size_t size)
 {
     return malloc(size);
 }
 
-void* operator new[](unsigned int size)
+void* operator new[](size_t size)
 {
     return malloc(size);
 }
